refactor: decompose proxy.rs (1981 LOC) into src/proxy/ module directory — PHASE 2
Pure structural refactoring — zero behavior changes:
- src/proxy/mod.rs (283 LOC) — proxy_handler dispatcher, validate_request, re-exports
- src/proxy/error_types.rs (82 LOC) — UpstreamError, parse_upstream_error, extract_nested_error
- src/proxy/rate_limit.rs (47 LOC) — L2 rate limit detection, constants, logging
- src/proxy/classify.rs (275 LOC) — ErrorClass, 3-layer classification, ErrorRegexes (OnceLock)
- src/proxy/discovery.rs (119 LOC) — ModelCapabilities, probe_model_limit, get_context_limit
- src/proxy/concurrency.rs (94 LOC) — ModelSemaphores, CircuitBreaker type, acquire_model_permit
- src/proxy/retry.rs (331 LOC) — resilient_send, resilient_send_raw with 3-layer retry
- src/proxy/non_streaming.rs (155 LOC) — handle_non_streaming, WebFetch interception loop
- src/proxy/streaming.rs (533 LOC) — handle_streaming, create_sse_stream, SSE event processing

77 tests pass, zero clippy warnings, cargo fmt clean.
